requirements.txt
requests
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
python-dotenv
apify-client
gspread
google-auth
openai
google-generativeai
opencv-python
